Task 2/requirements.txt
customtkinter>=5.2.2
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
langdetect>=1.0.9
deep-translator>=1.11.4
python-dotenv
soundfile
requests
openai-whisper
google-generativeai>=0.4.0